Preliminary 3.4 release : should be the same version as MERGE_TROPOMI_GOME_RING; make a copy of modified files instead of svn merge due to a lot of mismatch with the different versions FRM4DOAS/TROPOMI/GOME/DALIBOR since the beginning of 2018.  This version should be upload on git.
git-svn-id: http://svn-ae/ae_doas/qdoas/trunk/Src@1908 0551a244-8104-0410-bdc1-badfa85f4f5d

+++ engine/output_private.h
@@ -679,7 +679,7 @@ static inline void get_rms(struct output_field *this_field, double *rms, const E
   *rms = (!pTabFeno->rc)?pTabFeno->RMS:QDOAS_FILL_DOUBLE;
 }
 
-static inline void get_rms_calib(struct output_field *this_field, int *rms, const ENGINE_CONTEXT *pEngineContext __attribute__ ((unused)), int indexFenoColumn, int index_calib) {
+static inline void get_rms_calib(struct output_field *this_field, double *rms, const ENGINE_CONTEXT *pEngineContext __attribute__ ((unused)), int indexFenoColumn, int index_calib) {
   *rms = KURUCZ_buffers[indexFenoColumn].KuruczFeno[this_field->index_feno].rms[index_calib];
 }
